Quater3/Class-07-05-MAY-2024/assignment/todo-backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
mypy = "^1.10.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"

[tool.poetry.scripts]
dev= "todo_backend.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
